refactor: auto-fix 43 Clippy warnings across codebase
Applied cargo clippy --fix to resolve style and idiom warnings:
- Remove unused imports and variables
- Add Default implementations for new() methods
- Simplify boolean expressions and string operations
- Replace manual implementations with derive attributes
- Use idiomatic Rust patterns (matches!, if let, clamp)
- Fix redundant field names in struct initialization

Affected: 24 files across services, database, system_integration, task_queue, utils

+++ long-compress-assistant/src-tauri/src/database/models.rs
@@ -288,7 +288,7 @@ impl From<CompressionTaskDb> for crate::models::compression::CompressionTask {
             format,
             options,
             status,
-            progress: db.progress as f32,
+            progress: db.progress,
             created_at: db.created_at,
             started_at: db.started_at,
             completed_at: db.completed_at,
